# perf(config): change default png_compress_level to 1