## What type of PR is this? (check all applicable)

- [ ] Refactor
- [ ] Feature
- [ ] Bug Fix
- [x] Optimization
- [ ] Documentation Update
- [ ] Community Node Submission


## Have you discussed this change with the InvokeAI team?
- [x] Yes
- [ ] No, because:

## Description

This substantially reduces the time spent encoding PNGs. In workflows with many image outputs, this is a drastic improvement.

For a tiled upscaling workflow going from 512x512 to a scale factor of 4, this can provide over 15% speed increase.

---

There is a minor file size increase, but bytes are far cheaper than time.

Snippet from [profiling](https://github.com/psychedelicious/pil-vs-cv2):
file @ (1024, 1024), compress_level=1 -> 63.35ms +-0.08ms (1494.83kb)
file @ (1024, 1024), compress_level=6 -> 215.78ms +-0.30ms (1327.04kb)

`1` takes 30% of the time for a 13% larger filesize.

## Merge Plan

This PR can be merged when approved

<!--
A merge plan describes how this PR should be handled after it is approved.

Example merge plans:
- "This PR can be merged when approved"
- "This must be squash-merged when approved"
- "DO NOT MERGE - I will rebase and tidy commits before merging"
- "#dev-chat on discord needs to be advised of this change when it is merged"

A merge plan is particularly important for large PRs or PRs that touch the
database in any way.
-->

## [optional] Are there any post deployment tasks we need to perform?

@Millu We may want to communicate to users that they can set this in `invokeai.yaml` for a speed boost, in case the setting is already set for them.